Canvas demo: Adding drop-shadow to the groupedPin bubble
In the canvas demo, the bubble created when pins are grouped in the canvas, often blends in the background and doesn't look good. This PR adds a drop-shadow to the corresponding element, to make sure it separates from its surroundings.

The value I used is the same one as that of the relevant Figma element.

Test Plan:
The drop-shadow appears correctly on the bubble element.

https://github.com/getcord/monorepo/assets/102963329/7cecbd9d-0e40-48f7-b692-6cd0d2330808

<img width="340" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/b4d8a3b7-436f-4ae8-9720-ba92fd8c905c">


Reviewers: flooey, maylynn-ng

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7224

monorepo-commit: 0f13e490c86839ac4acdb425b9fe7787eb280359