Document demo: Removing "Connect to share to slack" button from the message options menu
We don't allow users of the demo apps to connect to Slack (at least for now). However there was a leftover "Connect to share" button which did nothing but was still present in the Message component options menu.

This PR hides this button via CSS.

Test Plan:
Tested in all components of the document demo that render a Message component (Thread and ThreadedComments). The button no longer appears in any of the two.

Thread component:

![Screenshot 2023-11-15 at 12 37 10](https://github.com/getcord/monorepo/assets/102963329/0017ad4d-258f-4cde-9cfb-092398061ee3)

ThreadedComments component:

![Screenshot 2023-11-15 at 12 37 17](https://github.com/getcord/monorepo/assets/102963329/3da1fb4b-1b0a-42ac-bfd3-e2d62c387383)


Reviewers: elenaperezrioja, flooey, maylynn-ng, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7263

monorepo-commit: 08341e32e8cadc29fa049030e92da8c96b83e378