demo-apps: canvas-new - facepile on pins
To match figma we want our pins to show not only the author of the first message, but of anyone who has sent a message within this thread!



Test Plan:
Reply to a thread and there should be a Facepile of all of the repliers & the author.
Threads with no replies / the replies are all from the author should only show the one user

Reviewers: elenaperezrioja, km-nur, flooey, nickfil22

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/7231

monorepo-commit: 9400b8cdd689feee418d388eb7c7b85bd9f15c09