Canvas demo: Upgrading Cord package
While implementing LiveCursors in the Canvas, we noticed that there was a case where
the Canvas may not be initialized yet, but the LiveCurors could be, so we didn't want to send
any mouse coordinates that were invalid. Turns out the `eventToLocation` function of the API
only allowed returning x/y coordinates. The new npm package version includes a null return type
on this function.

Test Plan:
eyes


Reviewers: flooey, elenaperezrioja, km-nur, maylynn-ng

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7310

monorepo-commit: 1000c22d10cd506455e45a14a6e1aea5420a525e