Canvas demo: Adding live cursors
A great place to demo the LiveCursors component is the Canvas, as we can really showcase the API, by mapping coordinates between different panning and scales on the canvases of other users.

This PR adds live cursors and correctly maps the coordinates between the sender and receiver end of the live cursors.

Test Plan:
In all test cases, to show that coordinates map correctly, I will be touching with the cursor, the bottom-left corner of the big arrow (approximately the center of the canvas), the top-right corner of the orange square and the bottom-right corner of the orange square. The cursor should be at the exact same point in both canvases for test cases 1-7.

1) Both canvases are in the same scale and same position
2) The sender canvas is shifted right
3) The receiver canvas is shifted down
4) The sender canvas is zoomed out
5) The receiver canvas is zoomed out
6) The receiver canvas is zoomed in
7) The sender canvas is zoomed in

*8)
Test case: Move the sender cursor in an area that should be outside the view of the receiver's canvas
Expected outcome: No cursor should appear in the receiver end

*9)
Test case: Move the sender cursor outside the canvas
Expected outcome: No cursor should appear in the receiver's end


https://github.com/getcord/monorepo/assets/102963329/8e680cb4-c446-4a54-8626-741d1b611ef1





Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7245

monorepo-commit: 63adca72e2a8fc2c4f2bfbff70b1c91ff8f2b48e