Canvas Demo: comments list: hover on thread should hide dividers
Quick update to make it look a bit nicer, but maybe not so nice-code.
As with these kind of PRs, it also includes a small change that fixes
a subtle border radius style on the comments list that looked a bit off.



### Test Plan:

Before:

https://github.com/getcord/monorepo/assets/127104309/9c723e04-0b92-41c5-bacc-6f00702ceb2c

After:

https://github.com/getcord/monorepo/assets/127104309/14cc2c01-1dd2-49bc-9fbb-6fc77f33d432

Reviewers: maylynn-ng, flooey, nickfil22, elenaperezrioja

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7257

monorepo-commit: 5e1c2626388d99e704450585541bda9fe30042ec