Canvas demo: Displaying resolved threads
When resolving a thread in the canvas demo, it completely disappears and cannot be recovered. For that reason, we want to add the ability to view resolved threads.

This PR adds tabs to the ThreadedComments component used in the canvas and styles them to match the rest of the canvas.

Test Plan:
Initially there are no resolved threads. The empty state has the expected resolved empty state text. After resolving a thread, it appears in the resolved tab. Then unresolving it, it goes back to the open tab. The empty state in both the Open and Resolved tab appear as expected.


https://github.com/getcord/monorepo/assets/102963329/55922be3-1bbf-4a47-944f-8f866c327c5f






Reviewers: km-nur, elenaperezrioja, maylynn-ng, flooey

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7238

monorepo-commit: dfaa33cb92f8372caac114206af37e21398bdf6d