QD: Adding page presence to the document demo 'how-its-built' dropdown
Looking at the 'How-its-built' dropdown for the document demo, I noticed that the PagePresence component was not mentioned, even though it is being used.

This PR adds the component name to the dropdown.

Test Plan:
After clicking on the dropdown, `Page Presence` was correctly displayed in the list. When clicking on the `Page Presence` option of the list, I was correctly redirected to the PagePresence docs page
<img width="244" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/0a9a82bf-3a20-41f0-bb55-26d811d0f59e">



Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7361

monorepo-commit: d1fe27b62d37b18c5e107c15aaa7d29e04afb438