Document demo: Making the opening of ThreadedComments smoother
The fact that ThreadedComments component in the document demo just suddenly appears when clicking the launcher button doesn't look great.

This PR makes the opening of the component smoother by modifying the floating component structure to allow adding transitions and also adding a transition to make the opening smoother.

Test Plan:
Video of opening and closing the component in the document demo


https://github.com/getcord/monorepo/assets/102963329/d56841bb-6bb5-42b5-95d7-abd8eafac8ed



Reviewers: maylynn-ng, elenaperezrioja, km-nur

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7360

monorepo-commit: d06b64e87e4c0a9f940a05e0d8ab0378297bf3b8