Document demo: Spatially sorting threads before displaying them
Floating threads in the document demo are sorted by timestamps initially, however, after the text animation is finished, they are re-sorted spatially, based on where in the text the highlight is. This causes the floating threads to shuffle around, which doesn't look great (see before video in test plan).

This PR moves the re-sorting of floating threads to the time they are rendered, making sure that they are always displayed in the correct order.

Test Plan:
Notice how the threads are re-shuffled once the animation finishes in the "before" video. Whereas in the "after" video, threads are in the correct order from the beginning.



Before:


https://github.com/getcord/monorepo/assets/102963329/0db055c3-a302-4c35-947e-9b9fbd916a76



After:

https://github.com/getcord/monorepo/assets/102963329/dfb6e50b-08c8-4ca2-a3a0-61175cd0497a









Reviewers: flooey, km-nur, maylynn-ng, elenaperezrioja

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7315

monorepo-commit: d1c202a81aefa7e6ce584b21f523369c6d008107