Canvas Demo: center thread on canvas when opening it
Originally, if the selected thread's pin was in a visible position
on the canvas, we'd open the thread at that position.
This diff removes that logic and centres the pin on the canvas
first before opening it.

Not sure if we'd rather have a smarter way of figuring out where in the
canvas the pin is and only centre it or move it if it's not within the
canvas's central area.
We've also now set it to close any open threads when doing the whole
navigation as the open thread would move around as the canvas shifted.


Test Plan:
Before:

https://github.com/getcord/monorepo/assets/127104309/0342c6c6-6aa2-42eb-a40f-f10d6cd4e7c1


After:

https://github.com/getcord/monorepo/assets/127104309/b6ca2dab-21e6-4d83-8bcd-411149b4a1c0

Reviewers: maylynn-ng, nickfil22, elenaperezrioja

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/7386

monorepo-commit: 6cca24960725b4ce11cd8c96b7b3764c8b6f21db