Dashboard demo: Fixing ThreadedComments empty state
The empty state of ThreadedComments in the dashboard demo jumped, as the empty state avatar was not aligned with the message avatar.

This PR styles the ThreadedComments empty state so that the avatar is the same size as the message avatar and is positioned in the same place as the avatar of the first open thread.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/5fc5810c-dc99-4116-a89b-c1e241e7ae79




Reviewers: elenaperezrioja, maylynn-ng, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7384

monorepo-commit: 3bd2dc3e1fb9035b77c407682e79af1de8ce02ea