Document demo: Fixing highlighting bug
When starting a selection within the sheet and we move the cursor outside it, the selection highlights a completely different part of the page (video of the bug below). This doesn't look great for the document demo app. This also happens when we're starting a selection from one of the floating threads and we move the cursor within the sheet.

We actually don't want to be able to highlight anywhere outside the sheet if we have started the highlight from the sheet and vice-versa.

This PR fixes this by:
1) Disabling selection in the areas outside the sheet when a highlight has started within the sheet
2) Disabling selection on the sheet, when a highlight has started outside the sheet.

Test Plan:
Before (Video of bug):


https://github.com/getcord/monorepo/assets/102963329/7e782bb6-2c6d-4ec2-8fcf-c041014e9c5f


After:

Case 1: Starting the selection from within the sheet, then moving the cursor outside
Outcome: The selection only highlights within the sheet

Case 2: Starting the selection from outside the sheet, going into the sheet
Outcome: No text is selected in the sheet


https://github.com/getcord/monorepo/assets/102963329/ce89cc8d-c23c-49e3-a120-65a9fbc9a33a





Reviewers: AlbertQM, km-nur, elenaperezrioja, maylynn-ng

Reviewed By: km-nur, maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7283

monorepo-commit: 99fd64f57e9d75b6ee651d11c7ed3932050219de