Demo apps: Highlighting components when hovering on 'How-its-built'
When a person in the wild is looking at the demo apps, they have no idea which part of them is Cord. In fact, they would have to find the 'how-it's-built' button, then click on it, then click on one of the components to see what it actually looks like in the documentation page. Then navigate back to the demo and find the component they saw in the docs within the demo (which may be styled differently than the docs, so it may not be clear which one it is).

This PR highlights each component on the demo when hovering over the relevant button in the 'How-its-built' dropdown. We can simply implement this with a context, as we have two separate React trees for the header and the body for the docs demo apps. So we achieve this by adding a data attribute to the root of the document which indicates which component is being hovered on. On the document demo side, we are adding a mutation observer in order to set the presence and show an avatar when a user is hovering on the `Avatar` entry of the dropdown.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/d9150a88-7fea-40bf-8fc2-8dce41c6b4c7





Reviewers: elenaperezrioja, maylynn-ng, flooey, km-nur

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7359

monorepo-commit: e0efedfa28a1c7708b73d757664d8d51afa353b7