Dashboard demo: Adding a quote with metadata on the first message of each thread
All dashboard demo comments are linked to either a point in a chart, or a cell in the grid. However, when looking at a thread out of context (e.g. in the ThreadedComments) we don't know what it refers to.

This PR improves this by adding a quote to the first message of each thread, with information about where this comment was made. This quote is only displayed on the ThreadedComments component, as it doesn't make sense to have it on the thread that is already in context.

Using the `beforeMessageCreate` callback, I'm adding a paragraph at the beginning of the first message of a thread and then styling it to look like a blockquote.

I am using a pseudoelement on the text instead of the `quote` node of the structured message because 1) the `quote` uses a `border-left` property, which does not allow us to style it with the necessary border-radius and 2) because we need to have a different colors based on which company the metadata refers to and this classname is stored on the element's child. So it makes more sense to directly style the child.

Test Plan:
Adding comments on the Notion and Figma chart lines, then on the grid. Notice that the quote doesn't appear on the threads, but does appear within the threads in ThreadedComments.

There is also a still image of the quotes in ThreadedComments below the video.


https://github.com/getcord/monorepo/assets/102963329/3271d8da-1ed7-4fc3-acce-6acb81834a6e

(Image)
<img width="305" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/0bbda9d7-c732-4df4-8ce8-b465538ce38c">





Reviewers: elenaperezrioja, km-nur, flooey, maylynn-ng

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7409

monorepo-commit: c9a441ff1f2755f8b3e55590878b78906e2a2f23