Canvas demo: allow clicking on the whole message in commentsList
Since the onMessageClick prop in threaded comments only account for the
message, we have an issue where clicking on the padding of the thread
containing the message deosn't dispatch the event. An easy way to get around
this on our side it to remove the thread padding and increase the message
paddig.


Test Plan:
Before:

https://github.com/getcord/monorepo/assets/127104309/c03f37ca-35ab-4fab-ab11-5ac8322db3a5

After:

https://github.com/getcord/monorepo/assets/127104309/875b697a-5cdc-4498-afa1-9a3021608889

Reviewers: nickfil22, maylynn-ng, elenaperezrioja

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7481

monorepo-commit: f53b84f98dbc82e2bdfd568009c1fe8d58d252fa