QD: Fixing repliers label on the Canvas demo
The text indicating how many replies there are in a Thread in the ThreadedComments component in the Canvas demo was stuck on the left border of the component. This styling is not what we want.

This PR adds a `padding-left` to the text so that it's aligned with the message component text right above it.

Test Plan:
Before:
![image](https://github.com/getcord/monorepo/assets/102963329/fa25fbb0-8a2c-42e9-9de9-802fdaf16b25)

After:
![image](https://github.com/getcord/monorepo/assets/102963329/80762b1f-b34c-414c-bbe0-632406922834)



Reviewers: km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7547

monorepo-commit: 0dd95cf7cc6e7b31123ee660aac5a13416c1552b