Dashboard demo: Highlighting components when hovering on 'How-its-built'
Same as #7359, but in the Dashboard demo

When a person in the wild is looking at the demo apps, they have no idea which part of them is Cord. In fact, they would have to find the 'how-it's-built' button, then click on it, then click on one of the components to see what it actually looks like in the documentation page. Then navigate back to the demo and find the component they saw in the docs within the demo (which may be styled differently than the docs, so it may not be clear which one it is).

This PR highlights each component on the demo when hovering over the relevant button in the 'How-its-built' dropdown.

Similarly to #7359, we still have to use a mutation observer to know when we are highlighting over the `Presence Facepile` and `Presence Observer` and set users present to show off the components.

I have slightly refactored how Thread and ThreadedComments are displayed, so that we can highlight these components using CSS only. For Thread, instead of conditionally changing the `visibility`, we now conditionally apply a classname, with which we change the `visibility`. In ThreadedComments, instead of conditionally rendering the component, we now conditionally apply a classname to its sibling button component and change the `visibility` CSS property.

When hovering over `Thread`, we open and highlight all threads. This is done intentionally because:
1) It is way cleaner to only use CSS (we would have to programmatically always bring the thread to open and highlight in view and it would be an additional clutter in our open-source code)
2) A user might resolve the specific thread to highlight

Test Plan:
Made sure that the Thread and ThreadedComments components still work correctly. Added new threads, opened them, closed them. Also opened and closed the ThreadedComments. All worked as expected.

Highlighting tested in the video below:

https://github.com/getcord/monorepo/assets/102963329/9a85ef65-5819-4656-8775-46b55197b1c5





Reviewers: km-nur, flooey, elenaperezrioja, maylynn-ng

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7432

monorepo-commit: 3e75091dce1929081c963a252169be069cc0dea1