Canvas demo: Highlighting components when hovering on 'How-its-built'
Same as #7359, but in the Canvas demo.

When a person in the wild is looking at the demo apps, they have no idea which part of them is Cord. In fact, they would have to find the 'how-it's-built' button, then click on it, then click on one of the components to see what it actually looks like in the documentation page. Then navigate back to the demo and find the component they saw in the docs within the demo (which may be styled differently than the docs, so it may not be clear which one it is).

This PR highlights each component on the demo when hovering over the relevant button in the 'How-its-built' dropdown. We only need CSS to apply the highlight for this demo!

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/a457aec4-fe01-4c74-9179-b954b61cb615

Screenshot of highlighted messages with and without repliers below:

<img width="426" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/187d5b46-cee2-4b6a-93d8-cd3663ad1eef">




Reviewers: maylynn-ng, km-nur, elenaperezrioja, flooey

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7494

monorepo-commit: 59875d49d23bed52a747e9112e032711df3ed34d