Video demo: Show resolve-thread button in ThreadedComments
In the video demo, we initially show ThreadedComments in the "unresolvedOnly"
threads state. However, when a user resolves a thread, we change the state to "tabbed",
so an `Open` and `Resolved` tab appear.

The issue with this is that when we are in the "unresolvedOnly" state, the "Resolve" button
is hidden in ThreadedComments. So a user can only resolve a thread from the threads attached
to a pin. This doesn't look great in terms of UI consistency.

This PR makes sure to always show the "Resolve" button on the ThreadedComments.

Test Plan:
Clicked on the options menu on the ThreadedComments and I could correctly see
the "Resolve" button. (Image below)

![image](https://github.com/getcord/monorepo/assets/102963329/89fb37c3-0160-465e-9321-660cd9577f80)



Reviewers: flooey, elenaperezrioja, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7573

monorepo-commit: 4a54b7bc4d338f279ab506a370813b2633041ae6