Video player: Bring comment into view when clicking on a notification
When clicking on a notification in the Video player demo, the whole page refreshed and we were redirected to the beginning of the video. This was not a very good experience.

This PR changes the behavior when clicking on notifications, to  redirect the user to the relevant thread that notification refers to, without refreshing the page at all. We already have a `onMessageFocus` function in the VideoPlayer demo, so we are simply reusing this one.

Test Plan:
Before and after videos of the behavior when a user clicks on a notification.

Before:

https://github.com/getcord/monorepo/assets/102963329/e02309ad-bc03-49f8-9d84-59731470e7dc

After:

https://github.com/getcord/monorepo/assets/102963329/e0e01334-28a7-4250-8161-e71f147c09b2




Reviewers: elenaperezrioja, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7602

monorepo-commit: 7c656a8d88201bdfe01abee428ad9d10288fb633