Canvas Demo: clicking on thread in list should close the open thread
In figma, when a thread is open and we click on the thread in comments list
it closes, rather than being re-opened as it is in our canvas.
This diff just changes that behaviour to work in a similar way.


Test Plan:
Before:

https://github.com/getcord/monorepo/assets/127104309/8f817b1c-99d3-47e8-a663-ae092e249ae1


After:

https://github.com/getcord/monorepo/assets/127104309/9dbb0b7b-17e1-4744-b20c-460409a02fe8

Reviewers: maylynn-ng, nickfil22, elenaperezrioja

Reviewed By: nickfil22

Pull Request: https://github.com/getcord/monorepo/pull/7482

monorepo-commit: 1da02646f0f6550850e21d3aa4a76af77df2466c