Dashboard demo bug fix on token refresh
When opening the dashboard demo in a state where the JW token needed refreshing, or when the local storage was cleared of a token, the demo appeared blank with an error in the console saying that the CordSDK was not initialized (image below). When refreshing the page everything worked as expected.

This happened because we initialized the CordSDK without a token, then populated the token, then the presence SDK was initialized - separately to the CordSDK object. So checking if `window.CordSDK` was initialized didn't help us know if the presence SDK was initialized.

This PR swaps the `window.CordSDK` null check in the dashboard demo with getting the `cordSDK` from the CordContext. This is a safer solution which will always return the correct result, as the CordContext will only be populated when the CordSDK is initialized.

Image of bug:
<img width="1000" alt="Untitled" src="https://github.com/getcord/monorepo/assets/102963329/6d0bcac6-484a-40f1-bde9-54561cb47754">

Test Plan:
Built the dashboard demo app. Cleared the token in the local docs, then refreshed the page. Everything loaded correctly.


Reviewers: km-nur, AlbertQM, elenaperezrioja

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7627

monorepo-commit: ee681bcb194c141f0f28b23034e169da21e0a33e