Stop demo apps from flashing when first opened
When first opening the demo apps there's a flash, meaning that the body background color is initially white, then changing to the actual background color of the respective demo app (Video attached). This doesn't look good. It is happening because our demo apps are built separately to the docs app. So once the docs demo app page loads in the client, it then fetches the js and css for the demo apps to populate the respective pages. So for a split second the demo app page is empty and is then populated with the demo app.

This PR fixes this behavior by setting the initial background color of the demo app page to be the same as that of the respective app. We unfortunately could not figure out a better way to fix this or a way to connect the colors in the docs page and the demo app css, so instead, I have added a comment in the duplicated colors, to warn developers who might want to change it.

Note: The canvas app has a white background, so there is no need to modify it.
Note 2: I did consider to also add a comment in each demo app css file, but because they are in open-source repositories, I didn't want to have that comment in there.

Test Plan:
Video of bug:

https://github.com/getcord/monorepo/assets/102963329/491dfb7c-c10b-45e0-a14e-06974efd23e8


Same pages after the fix:

https://github.com/getcord/monorepo/assets/102963329/a5dabae2-0fdc-40ec-a614-ae7b76039f3f





Reviewers: elenaperezrioja, km-nur, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7609

monorepo-commit: b8818770253342a21517934871252c3b4a5edaf4