Canvas Demo: make message preview transitions smoother
The transitions were a bit janky when we had longer text in the message.
The text would wrap as it grew to fill the container so didn't look very
neat or consistent. The problem wasn't there when the container shrunk
back to the pin-size as we had some css in place to not dispay the message
once we stopped hovering.

This diff cleans up the behaviour slightly buy making the message text
only appear once the container transition is done.


Test Plan:
**Before:**

https://github.com/getcord/monorepo/assets/127104309/5d3be9cc-82d0-40de-b55e-49c663b01a4a

**After:**

https://github.com/getcord/monorepo/assets/127104309/89d550f4-a174-443d-94da-718d864e7267

Reviewers: nickfil22, elenaperezrioja

Reviewed By: nickfil22

Pull Request: https://github.com/getcord/monorepo/pull/7595

monorepo-commit: afff2dd58cf922c71f6ca8f0e36aab75ab9a482a