Video demo: Highlighting components when hovering on 'How-its-built'
Same as #7359, but in the Video demo.

When a person in the wild is looking at the demo apps, they have no idea which part of them is Cord. In fact, they would have to find the 'how-it's-built' button, then click on it, then click on one of the components to see what it actually looks like in the documentation page. Then navigate back to the demo and find the component they saw in the docs within the demo (which may be styled differently than the docs, so it may not be clear which one it is).

This PR highlights each component on the demo when hovering over the relevant button in the 'How-its-built' dropdown.

I have slightly refactored how the Thread component is displayed so that we can highlight it with css only. Instead of conditionally changing the `display` css property, we now conditionally apply a classname and change the `display` css property based on that class.

Because the odds of a pin not being in view when a user highlights the `Pin` or `Thread` component are quite high in this demo, I made a decision to bring them into view programmatically. I think it's important to properly showcase these components, even though it's a lot of code which is not useful for developers.

I also abstracted the code which checks whether a pin timestamp is within 1.5seconds of the current timestamp, so it can be re-used outside to see if a pin is in view.

Test Plan:
Testing the highlighting when we already have a pin in view and when we don't have one in view. The former just highlights components as expected, whereas the latter brings a pin into view when attempting to highlight it.

https://github.com/getcord/monorepo/assets/102963329/a2cbb8fa-c85d-45ee-a8a1-43787cce550f





Reviewers: flooey, elenaperezrioja, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7576

monorepo-commit: 1bf5f2a6481add0a5535a6ab6dc9275cb8c72342