Document demo: Fixing reply from email icon stuck on username
When replying via email in a comment made on the document demo, the `sent via email` icon of the message component was stuck on the username (See "before" image below).

This was happening because of a combination of two factors.
1) There was no margin between the username and the "sent via" icons
2) The document demo uses a modified message component grid layout, which has the username and "sent via" icons on the first line, and the timestamp right below. Because the timestamp was spanning 2 grid template columns - the same ones as the username and "sent via" icons respectively, when its text was short ("just now"), it squeezed the two icons together and when it had longer text, such as "1 minute ago",  it would pull the username and icon away from each other.

This PR fixes this by allowing the timestamp to extend three columns wide, so it doesn't affect the username and icons and by adding a margin between the username and "sent via" icon.

Test Plan:
Before:
![image](https://github.com/getcord/monorepo/assets/102963329/21d0b1d8-459d-4061-bcf4-700c3b89a0bf)
![image](https://github.com/getcord/monorepo/assets/102963329/6a2a4a44-e2bb-4f9c-8aba-829617fda632)


After:
![image](https://github.com/getcord/monorepo/assets/102963329/17b2b4d2-75dd-432a-aed3-1fcb4502b8f1)
![image](https://github.com/getcord/monorepo/assets/102963329/98cdc611-ef98-4531-a6bd-cc485b6d3de5)




Reviewers: elenaperezrioja, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7640

monorepo-commit: efa7a197534d9d881ebb4f7675792db83e90bce1