demo-apps: rename dashboard-new to dashboard
This PR renames the folder from `dashboard-new` to `dashboard`.
It also renames the urls from `dashboard-new` and `document-new` to just `dashboard` and `document` and adds a redirect for these two.

What I'm a little unsure of is how this works with cord.com? I don't think this would affect that - is that correct?

Still to do:
update the repo (then maybe that will affect cord.com?)


Test Plan:
Everything compiles
You can navigate to `/document` or `/dashboard`, if you navigate to either of these `-new` it will redirect you to `/document` or `/dashboard`

Reviewers: nickfil22, km-nur, jwatzman, elenaperezrioja

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7528

monorepo-commit: 3039133b0f464ba9abe0a3281eca056aa0497a1c