demo-apps: video player - align resolved/unresolved tabs to the top of video
Thanks @mhtien for spotting this! :pray:

This PR aligns the resolved/unresolved tab container to the top of the video player when it's at it's max width.

Making the window smaller will shrink the video, but the threaded comments stays the same size so the tabs don't move either.

**before**

<img width="399" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/1e9e9f86-09dc-492c-b483-fe34ee377867">


**after**

<img width="442" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/de6586bc-666a-4f7f-81ea-5cd94d78b257">



Test Plan: Have a look

Reviewers: mhtien, km-nur, nickfil22

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7715

monorepo-commit: 69c1934aec2225815a05fbe7976daa35535b19ca