demo-apps: video player - notification options menu styling
Currently you can't really see the options menu when hovering on a notification because the colour is too similar to it's background. Also the hover state the colour is too close to it's background.

This PR changes the colour for the options button text

**before**

https://github.com/getcord/monorepo/assets/62358728/f39491f2-52e8-4b96-b06a-35e62c7c16f4


**after**

https://github.com/getcord/monorepo/assets/62358728/a3b0bb02-ebac-4655-aed2-76a47ebc4b62


Test Plan: Have a look

Reviewers: mhtien, km-nur, nickfil22

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7717

monorepo-commit: 85be68a780fc7efa168dfcd8b4855dabdbefe270