Demo-app: always take screenshot on new-message and new-thread
There was a bug where screenshots weren't being created for new threads -
meaning if a notification was from sending the first message in a thread
(which is what creates the thread) then we don't create a screenshot.

This also includes bumping the cord package to 1.30.1 to include a syntax
fix for the screenshot config to make it all work.


Test Plan:
Before:
Screenshot not being taken on new thread:

https://github.com/getcord/monorepo/assets/127104309/d428e30a-3cab-435e-b10d-a399d6f47c11


After:

Screenshot captured on new thread and new message:

https://github.com/getcord/monorepo/assets/127104309/05fa43a7-a1b3-44c6-a36b-78837534eae0

https://github.com/getcord/monorepo/assets/127104309/b363795d-d481-4a84-a82b-9bd50821eb85

<img width="467" alt="Screenshot 2024-01-08 at 11 35 31" src="https://github.com/getcord/monorepo/assets/127104309/b3616835-b3e2-4522-8232-4b74a7a0387a">

Reviewers: AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7755

monorepo-commit: 6062f252843fcde92b09ec65aa77e6f7e0fed427