Demo app dropdown redesign (1/3): Refactoring the header
We want to implement the new designs for demo app "How it's built" dropdown. To do so, we need to re-frame each demo app (before and after framing images below) and then add the new dropdown - which we can now call "drawer". I have broken this down into 3 main steps and a stack of PRs, in order to make this easier to review.

The first step (this PR) removes the blur and background from the demo app header and positions it absolutely. We are removing the background, as our new designs have 2 main background themes, light and dark. So we just want the header to be transparent. The change to `position: absolute` happens because the new drawer has a lot of elements, that overlaying the demo app content, makes things too busy. So we don't want the header to ever overlay the content.

Splitting the background and positioning changes into the first step, allows us to make the necessary changes in the styling of each demo app separately, without the apps looking really bad or mixing the changes with the new drawer PR ones.

Before framing:
<img width="1142" alt="Screenshot 2023-12-20 at 17 57 55" src="https://github.com/getcord/monorepo/assets/102963329/fd9779df-997b-4bae-81e4-bcf3bc82a090">
<img width="1138" alt="Screenshot 2023-12-20 at 17 58 04" src="https://github.com/getcord/monorepo/assets/102963329/a52724c4-e9f9-4d06-ba5e-ba70adb167fb">


After framing:
<img width="586" alt="Screenshot 2023-12-20 at 17 47 07" src="https://github.com/getcord/monorepo/assets/102963329/d7b91997-46b3-41b5-8acb-d4c3c0065e22">
<img width="584" alt="Screenshot 2023-12-20 at 17 47 14" src="https://github.com/getcord/monorepo/assets/102963329/86494c7d-1e1e-446c-a1bd-297241480bf2">


The second step will make the necessary styling and layout changes to each demo app to make them look like the designs in the "After framing" images above.

The third step will introduce the drawer.

Test Plan:
First tested the demo apps locally and then in the local docs. Notice the header blending with the rest of the app and the header scrolling out of view.

https://github.com/getcord/monorepo/assets/102963329/b908bd43-f694-4821-b333-0d33c9617eca



Reviewers: elenaperezrioja, km-nur, maylynn-ng

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7691

monorepo-commit: c1af678cc926efeb8ae3f60b9966959542d52ebf