Demo app dropdown redesign (2.1/3): Framing the dashboard demo
Continuation of the plan outlined in: #7691

To implement the new "How it's built" dropdown we need to change the designs of the demo apps and frame them.

This PR implements the new designs for the Dashboard demo by changing the styles of the relevant container divs.

Test Plan:
The screenshot below shows the demo app. It looks correct, same as the designs.

Also tested with the following headers used on the website: `header=false&website-demo=true`, the demo looks the same as the production docs - so no change will be visible in cord.com
<img width="1510" alt="Screenshot 2023-12-20 at 18 18 41" src="https://github.com/getcord/monorepo/assets/102963329/5ebb9104-b2da-4cfd-baca-afc7e6354558">


Reviewers: maylynn-ng, km-nur, elenaperezrioja

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7692

monorepo-commit: 01ea9acc68d199377d4d95a86472d70622e86ad1