Demo app dropdown redesign (2.2/3): Framing the document demo
Continuation of the plan outlined in: https://github.com/getcord/monorepo/pull/7691

To implement the new "How it's built" dropdown we need to change the designs of the demo apps and frame them.

This PR implements the new designs for the Document demo by changing the styles of the relevant container divs. I had to add a wrapper div to the document div, so I can nicely frame it, which also meant making it the infinite scroll container.

Test Plan:
The screenshot below shows the demo app. It looks correct, same as the designs. Also tested adding comments, scrolling down and making sure that the demo app behaves the exact same way as before.

Also tested with the following headers used on the website: `header=false&website-demo=true`, the demo looks the same as the production docs - so no change will be visible in cord.com

<img width="1510" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/56355c21-6777-4bbf-90e8-f88cdeda2e87">


Reviewers: km-nur, maylynn-ng, elenaperezrioja

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7696

monorepo-commit: e6407cf19866981a98de935f13493fcf47928512