Demo app dropdown redesign (2.3/3): Framing the video demo
Continuation of the plan outlined in: https://github.com/getcord/monorepo/pull/7691

To implement the new "How it's built" dropdown we need to change the designs of the demo apps and frame them.

This PR implements the new designs for the Video demo by changing the styles of the relevant container divs.

Test Plan:
The screenshot below shows the demo app. It looks correct.

Also tested with the following headers used on the website: `header=false&website-demo=true`, the demo looks the same as the production docs - so no change will be visible in cord.com
<img width="1512" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/9314029c-9ae1-4692-ba9a-0cceebfb9eb0">

Reviewers: km-nur, maylynn-ng, elenaperezrioja

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7697

monorepo-commit: 405def88403b0b958f9be92258f5c38737f0dba9