Demo app dropdown redesign (2.4/3): Framing the canvas demo
Continuation of the plan outlined in: https://github.com/getcord/monorepo/pull/7691

To implement the new "How it's built" dropdown we need to change the designs of the demo apps and frame them.

This PR implements the new designs for the Canvas demo. In this case we only need to change the background of the demo from white to a light grey. The demo is otherwise framed already.

Test Plan:
The screenshot below shows the demo app. It looks correct.

Also tested with the following headers used on the website: `header=false&website-demo=true`, the demo looks the same as the production docs - so no change will be visible in cord.com
<img width="1511" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/3c68973b-7376-423d-8da0-25ad5dc98cb4">


Reviewers: elenaperezrioja, km-nur, maylynn-ng

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7699

monorepo-commit: f07771c6401216f2351ba0210bbd37534ae328fe