Document demo: Open threads always appear over header
In the document demo, when creating a new comment on the top of the document and opening the mention list, the list appeared behind the header (See "before" image below). This happened because the z-index of the open thread was lower than that of the header, so even though the z-index of the mention list was the max-z-index, it was under the header in the stacking context. This bug also occurs with the reaction box.

This PR fixes this by setting the open thread z-index to be just 1 point larger than that of the header, to ensure that the stacking context of the open thread is on top of the header when present. Because the ThreadedComments component belongs to the stacking context of the header, it also ends up appearing behind the open thread. This PR ensures that we also close any open threads when clicking to open the ThreadedComments component, in order to avoid this UI state.

Test Plan:
Before:
<img width="362" alt="Screenshot 2023-12-19 at 12 03 03" src="https://github.com/getcord/monorepo/assets/102963329/d9b089a7-c879-4597-b901-2fba8b7305f6">

After:
<img width="400" alt="Screenshot 2023-12-19 at 12 01 50" src="https://github.com/getcord/monorepo/assets/102963329/ddbeb969-b041-4fd0-ae02-c0a5aa4d5480">

Also tested opening the ThreadedComments while having an open thread. The open thread went out of focus, then the ThreadedComments appeared on top of all threads.


Reviewers: km-nur, maylynn-ng, elenaperezrioja

Reviewed By: km-nur, maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7813

monorepo-commit: 5c1f7cf982382cc8b27d14cf9e850ba37fdbc722